Refactor constants module and update imports
Moved constants.py to utils/constants.py and updated all relevant imports in analyzer modules to reflect the new location. Added EMBEDDINGS_PATH to constants for future use.
