+++ src/utils/constants.py
@@ -8,3 +8,4 @@ PROJECT_ROOT = pathlib.Path(__file__).parent.parent
 DATA_PATH = PROJECT_ROOT / "data"
 EXAMPLE_CIPHERS_PATH = DATA_PATH / "example_ciphers"
 FREQUENCIES_PATH = DATA_PATH / "frequencies"
+EMBEDDINGS_PATH = DATA_PATH / "embeddings"
